concord-common: extract ZipUtils from IOUtils
