+++ cli/src/main/java/com/walmartlabs/concord/cli/RemoteRun.java
@@ -28,9 +28,9 @@ import com.walmartlabs.concord.cli.secrets.CliSecretService;
 import com.walmartlabs.concord.client2.ApiClient;
 import com.walmartlabs.concord.client2.ApiException;
 import com.walmartlabs.concord.client2.ProcessApi;
-import com.walmartlabs.concord.common.IOUtils;
 import com.walmartlabs.concord.common.PathUtils;
 import com.walmartlabs.concord.common.TemporaryPath;
+import com.walmartlabs.concord.common.ZipUtils;
 import com.walmartlabs.concord.sdk.Constants;
 import org.apache.commons.compress.archivers.zip.ZipArchiveOutputStream;
 import picocli.CommandLine.Command;
@@ -195,9 +195,9 @@ public class RemoteRun implements Callable<Integer> {
 
         try (var zip = new ZipArchiveOutputStream(dst.path(), StandardOpenOption.WRITE, StandardOpenOption.TRUNCATE_EXISTING)) {
             if (Files.isRegularFile(src)) {
-                IOUtils.zipFile(zip, src, src.getFileName().toString());
+                ZipUtils.zipFile(zip, src, src.getFileName().toString());
             } else if (Files.isDirectory(src)) {
-                IOUtils.zip(zip, src, PAYLOAD_ARCHIVE_FILTERS);
+                ZipUtils.zip(zip, src, PAYLOAD_ARCHIVE_FILTERS);
             }
         }
