+++ common/src/test/java/com/walmartlabs/concord/common/IOUtilsTest.java
@@ -20,7 +20,6 @@ package com.walmartlabs.concord.common;
  * =====
  */
 
-import org.apache.commons.compress.archivers.zip.ZipArchiveOutputStream;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
@@ -33,29 +32,6 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class IOUtilsTest {
-
-    @Test
-    public void testZipUnzip() throws Exception {
-        Path src = Files.createTempDirectory("test-zip");
-        Files.createFile(src.resolve("a.txt"));
-        Files.createFile(src.resolve("b\\c.txt"));
-        Files.createDirectory(src.resolve("b"));
-        Files.createFile(src.resolve("b").resolve("c.txt"));
-
-        Path archive = Files.createTempFile("archive", "zip");
-
-        try (ZipArchiveOutputStream zip = new ZipArchiveOutputStream(Files.newOutputStream(archive))) {
-            IOUtils.zip(zip, src);
-        }
-
-        PathUtils.deleteRecursively(src);
-
-        Path dst = Files.createTempDirectory("test");
-        IOUtils.unzip(archive, dst);
-        assertTrue(Files.exists(dst.resolve("a.txt")));
-        assertTrue(Files.exists(dst.resolve("b\\c.txt")));
-        assertTrue(Files.exists(dst.resolve("b").resolve("c.txt")));
-    }
 
     @Test
     public void testCopy() throws Exception {
