+++ it/server/src/test/java/com/walmartlabs/concord/it/server/MultipleProjectFilesIT.java
@@ -23,8 +23,8 @@ package com.walmartlabs.concord.it.server;
 import com.walmartlabs.concord.client2.ProcessApi;
 import com.walmartlabs.concord.client2.ProcessEntry;
 import com.walmartlabs.concord.client2.StartProcessResponse;
-import com.walmartlabs.concord.common.IOUtils;
 import com.walmartlabs.concord.common.PathUtils;
+import com.walmartlabs.concord.common.ZipUtils;
 import org.apache.commons.compress.archivers.zip.ZipArchiveOutputStream;
 import org.junit.jupiter.api.Test;
 
@@ -72,7 +72,7 @@ public class MultipleProjectFilesIT extends AbstractServerIT {
     private static Path zip(Path src) throws IOException {
         Path dst = PathUtils.createTempFile("template", ".zip");
         try (ZipArchiveOutputStream zip = new ZipArchiveOutputStream(Files.newOutputStream(dst))) {
-            IOUtils.zip(zip, src);
+            ZipUtils.zip(zip, src);
         }
 
         Set<PosixFilePermission> s = new HashSet<>();
