+++ it/server/src/test/java/com/walmartlabs/concord/it/server/ProjectFileIT.java
@@ -21,7 +21,7 @@ package com.walmartlabs.concord.it.server;
  */
 
 import com.walmartlabs.concord.client2.*;
-import com.walmartlabs.concord.common.IOUtils;
+import com.walmartlabs.concord.common.ZipUtils;
 import com.walmartlabs.concord.sdk.Constants;
 import org.apache.commons.compress.archivers.zip.ZipArchiveOutputStream;
 import org.junit.jupiter.api.Test;
@@ -112,8 +112,8 @@ public class ProjectFileIT extends AbstractServerIT {
         Path src = Paths.get(DependenciesIT.class.getResource("projectfile/deps").toURI());
         ByteArrayOutputStream baos = new ByteArrayOutputStream();
         try (ZipArchiveOutputStream zip = new ZipArchiveOutputStream(baos)) {
-            IOUtils.zip(zip, src);
-            IOUtils.zip(zip, tmpDir);
+            ZipUtils.zip(zip, src);
+            ZipUtils.zip(zip, tmpDir);
         }
 
         byte[] payload = baos.toByteArray();
