+++ plugins/tasks/concord/src/main/java/com/walmartlabs/concord/client/ConcordTask.java
@@ -22,8 +22,8 @@ package com.walmartlabs.concord.client;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.walmartlabs.concord.client2.*;
-import com.walmartlabs.concord.common.IOUtils;
 import com.walmartlabs.concord.common.PathUtils;
+import com.walmartlabs.concord.common.ZipUtils;
 import com.walmartlabs.concord.sdk.*;
 import org.apache.commons.compress.archivers.zip.ZipArchiveOutputStream;
 import org.slf4j.Logger;
@@ -766,7 +766,7 @@ public class ConcordTask extends AbstractConcordTask {
         if (Files.isDirectory(path)) {
             Path tmp = PathUtils.createTempFile("payload", ".zip");
             try (ZipArchiveOutputStream out = new ZipArchiveOutputStream(Files.newOutputStream(tmp))) {
-                IOUtils.zip(out, path);
+                ZipUtils.zip(out, path);
             }
             return tmp;
         }
