+++ plugins/tasks/concord/src/main/java/com/walmartlabs/concord/client/ConcordTaskCommon.java
@@ -22,8 +22,8 @@ package com.walmartlabs.concord.client;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.walmartlabs.concord.client2.*;
-import com.walmartlabs.concord.common.IOUtils;
 import com.walmartlabs.concord.common.PathUtils;
+import com.walmartlabs.concord.common.ZipUtils;
 import com.walmartlabs.concord.runtime.v2.sdk.TaskResult;
 import com.walmartlabs.concord.runtime.v2.sdk.UserDefinedException;
 import com.walmartlabs.concord.sdk.Constants;
@@ -637,7 +637,7 @@ public class ConcordTaskCommon {
         if (Files.isDirectory(path)) {
             Path tmp = PathUtils.createTempFile("payload", ".zip");
             try (ZipArchiveOutputStream out = new ZipArchiveOutputStream(Files.newOutputStream(tmp))) {
-                IOUtils.zip(out, path);
+                ZipUtils.zip(out, path);
             }
             return tmp;
         }
